add middleware for article route.

+++ client/src/components/ArticleDetails.js
@@ -1,16 +1,25 @@
 import React from "react";
 
 import { useArticlesContext } from "../hooks/useArticlesContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import { format } from "date-fns";
 
 const ArticleDetails = ({ article }) => {
   const { dispatch } = useArticlesContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      console.log("must be logged in to delete.");
+      return;
+    }
     const response = await fetch("/api/articles/" + article._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
 
     const json = await response.json();
